LPS-73034 Move the upgrade process to the proper module since resource actions for calendar are defined in calendar-service
