+++ modules/apps/forms-and-workflow/calendar/calendar-service/src/main/java/com/liferay/calendar/internal/upgrade/CalendarServiceUpgrade.java
@@ -18,10 +18,15 @@ import com.liferay.calendar.internal.upgrade.v1_0_4.UpgradeClassNames;
 import com.liferay.calendar.internal.upgrade.v1_0_5.UpgradeCalendarResource;
 import com.liferay.calendar.internal.upgrade.v1_0_5.UpgradeCompanyId;
 import com.liferay.calendar.internal.upgrade.v1_0_5.UpgradeLastPublishDate;
+import com.liferay.calendar.internal.upgrade.v1_0_6.UpgradeResourcePermissions;
+import com.liferay.calendar.service.CalendarResourceLocalService;
 import com.liferay.calendar.upgrade.v1_0_2.UpgradeCalendar;
-import com.liferay.calendar.upgrade.v1_0_6.UpgradeSchema;
+import com.liferay.calendar.upgrade.v2_0_0.UpgradeSchema;
 import com.liferay.portal.kernel.service.ClassNameLocalService;
 import com.liferay.portal.kernel.service.CompanyLocalService;
+import com.liferay.portal.kernel.service.ResourceActionLocalService;
+import com.liferay.portal.kernel.service.ResourceBlockLocalService;
+import com.liferay.portal.kernel.service.RoleLocalService;
 import com.liferay.portal.kernel.service.UserLocalService;
 import com.liferay.portal.kernel.upgrade.DummyUpgradeStep;
 import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
@@ -71,8 +76,21 @@ public class CalendarServiceUpgrade implements UpgradeStepRegistrator {
 			new UpgradeCompanyId(), new UpgradeLastPublishDate());
 
 		registry.register(
-			"com.liferay.calendar.service", "1.0.5", "1.0.6",
+			"com.liferay.calendar.web", "1.0.5", "1.0.6",
+			new UpgradeResourcePermissions(
+				_calendarResourceLocalService, _resourceActionLocalService,
+				_resourceBlockLocalService, _roleLocalService));
+
+		registry.register(
+			"com.liferay.calendar.service", "1.0.6", "2.0.0",
 			new UpgradeSchema());
+	}
+
+	@Reference(unbind = "-")
+	protected void setCalendarResourceLocalService(
+		CalendarResourceLocalService calendarResourceLocalService) {
+
+		_calendarResourceLocalService = calendarResourceLocalService;
 	}
 
 	@Reference(unbind = "-")
@@ -90,12 +108,35 @@ public class CalendarServiceUpgrade implements UpgradeStepRegistrator {
 	}
 
 	@Reference(unbind = "-")
+	protected void setResourceActionLocalService(
+		ResourceActionLocalService resourceActionLocalService) {
+
+		_resourceActionLocalService = resourceActionLocalService;
+	}
+
+	@Reference(unbind = "-")
+	protected void setResourceBlockLocalService(
+		ResourceBlockLocalService resourceBlockLocalService) {
+
+		_resourceBlockLocalService = resourceBlockLocalService;
+	}
+
+	@Reference(unbind = "-")
+	protected void setRoleLocalService(RoleLocalService roleLocalService) {
+		_roleLocalService = roleLocalService;
+	}
+
+	@Reference(unbind = "-")
 	protected void setUserLocalService(UserLocalService userLocalService) {
 		_userLocalService = userLocalService;
 	}
 
+	private CalendarResourceLocalService _calendarResourceLocalService;
 	private ClassNameLocalService _classNameLocalService;
 	private CompanyLocalService _companyLocalService;
+	private ResourceActionLocalService _resourceActionLocalService;
+	private ResourceBlockLocalService _resourceBlockLocalService;
+	private RoleLocalService _roleLocalService;
 	private UserLocalService _userLocalService;
 
 }
