+++ modules/apps/forms-and-workflow/calendar/calendar-service/src/main/resources/com/liferay/calendar/upgrade/v2_0_0/dependencies/update.sql
