+++ modules/apps/forms-and-workflow/calendar/calendar-test/src/testIntegration/java/com/liferay/calendar/upgrade/v2_0_0/test/UpgradeSchemaTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.calendar.upgrade.v1_0_6.test;
+package com.liferay.calendar.upgrade.v2_0_0.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.calendar.model.Calendar;
@@ -21,7 +21,7 @@ import com.liferay.calendar.model.CalendarBookingConstants;
 import com.liferay.calendar.model.CalendarResource;
 import com.liferay.calendar.service.CalendarBookingLocalService;
 import com.liferay.calendar.service.CalendarBookingLocalServiceUtil;
-import com.liferay.calendar.upgrade.v1_0_6.UpgradeSchema;
+import com.liferay.calendar.upgrade.v2_0_0.UpgradeSchema;
 import com.liferay.calendar.util.CalendarResourceUtil;
 import com.liferay.calendar.util.CalendarUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
