+++ modules/apps/forms-and-workflow/calendar/calendar-web/src/main/java/com/liferay/calendar/web/internal/upgrade/CalendarWebUpgrade.java
@@ -17,7 +17,6 @@ package com.liferay.calendar.web.internal.upgrade;
 import com.liferay.calendar.service.CalendarResourceLocalService;
 import com.liferay.calendar.web.internal.upgrade.v1_0_0.UpgradePortletId;
 import com.liferay.calendar.web.internal.upgrade.v1_0_0.UpgradePortletPreferences;
-import com.liferay.calendar.web.internal.upgrade.v1_0_2.UpgradeResourcePermissions;
 import com.liferay.calendar.web.internal.upgrade.v1_1_1.UpgradeEventsDisplayPortletId;
 import com.liferay.portal.kernel.service.PortletPreferencesLocalService;
 import com.liferay.portal.kernel.service.ResourceActionLocalService;
@@ -56,13 +55,7 @@ public class CalendarWebUpgrade implements UpgradeStepRegistrator {
 				UpgradePortletPreferences());
 
 		registry.register(
-			"com.liferay.calendar.web", "1.0.1", "1.0.2",
-			new UpgradeResourcePermissions(
-				_calendarResourceLocalService, _resourceActionLocalService,
-				_resourceBlockLocalService, _roleLocalService));
-
-		registry.register(
-			"com.liferay.calendar.web", "1.0.2", "1.1.0",
+			"com.liferay.calendar.web", "1.0.1", "1.1.0",
 			new com.liferay.calendar.web.internal.upgrade.v1_1_0.
 				UpgradePortletId());
